pyproject.toml
[tool.poetry]
name = "llm-toy"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "llm_toy"}]

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.27.4"
pyyaml = "^6.0"
streamlit = "^1.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
